restaurant_management/restaurant_management/requirements.txt
Django==5.1.2
djangorestframework
psycopg2-binary
python-dateutil
pytest pytest-django
factory-boy
injector